requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
coverage==7.6.1
Django==5.1.1
django-browser-reload==1.15.0
django-tailwind==3.8.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
wsproto==1.2.0
